+++ lib/channels/index.ts
@@ -2,4 +2,3 @@ export * from "./types";
 export * from "./errors";
 export * from "./registry";
 export * from "./service";
-export * from "./workflow";